+++ dunamai/__main__.py
@@ -66,7 +66,8 @@ common_sub_args = [
         "help": (
             "Custom output format. Available substitutions:"
             " {base}, {stage}, {revision}, {distance}, {commit}, {dirty},"
-            " {tagged_metadata}, {epoch}, {branch}, {branch_escaped}, {timestamp}"
+            " {tagged_metadata}, {epoch}, {branch}, {branch_escaped}, {timestamp},"
+            " {major}, {minor}, {patch}"
         ),
     },
     {